Simple setup script so that a database can get some dummy data
Small fix

Auth + improvements

Some fixes

Lots of fixes

Working on the admin command

About to run cargo fix

Output of cargo fix

renamed to admin-cli

Running cargo fix again

post-cargo-fix

End of round logic (still have to determine winner rewards)

Removed debug

Renamed to match rust naming conventions
