+++ spacetime-module/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "spacetime-module"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+serde = { version = "1.0.136", features = ["derive"] }
+spacetimedb = "0.3.0"
+rand = { version = "0.8.5", default-features = false }
+rand_chacha = "0.3.1"
+getrandom = { version = "0.2", features = ["custom"] }
